requirements.txt
about-time==4.2.1
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alive-progress==3.1.5
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
cattrs==23.2.3
cffi==1.16.0
cma==3.2.2
colorama==0.4.6
colour-demosaicing==0.2.5
colour-science==0.4.4
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
Deprecated==1.2.14
diffusers==0.30.0
dill==0.3.8
distro==1.9.0
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
facexlib==0.3.0
filterpy==1.4.5
fonttools==4.51.0
Forward_Warp==0.0.1
# forward_warp_cuda==0.0.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
grapheme==0.6.0
grpcio==1.63.0
huggingface-hub==0.24.5
icecream==2.1.3
imageio==2.34.1
imageio-ffmpeg==0.5.1
imgaug==0.4.0
importlib_metadata==7.1.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightning-utilities==0.11.5
llvmlite==0.42.0
lmdb==1.4.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
meshzoo
mkl-service==2.4.0
mmengine==0.10.4
moviepy==1.0.3
multidict==6.0.5
nest-asyncio==1.6.0
ninja==1.11.1.1
numba==0.59.1
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
portalocker==2.10.1
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-machineid==0.6.0
pycocotools==2.0.7
pycparser==2.22
Pygments==2.18.0
pyiqa==0.1.11
pymoo==0.6.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-package-info
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rawpy==0.21.0
regex==2024.5.10
requests-cache==1.2.1
rich==13.7.1
rich-argparse==1.5.2
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.4
six==1.16.0
sk-video==1.1.10
stack-data==0.6.3
stonefish-license-manager
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.5.10
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.3
torchaudio==2.0.2
torchmetrics==1.4.0.post0
torchvision==0.15.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.37.2
triton==2.0.0
tzdata==2024.1
ultralytics==8.2.16
url-normalize==1.4.3
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
x21==0.5.2
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
